backend/requirements.txt
# Python Backend Requirements
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg[binary]==3.2.3
python-decouple==3.8
Pillow==11.0.0
reportlab==4.2.5
